requirements.txt
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
idna==2.7
jsonlines==1.2.0
kiwisolver==1.0.1
matplotlib==3.0.0
numpy==1.15.1
pandas==0.23.4
pdfminer.six==20191110
pycryptodome==3.6.6
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
sortedcontainers==2.1.0
urllib3==1.23
